@@ -12,12 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from . import core
 from . import lazy
 from .interpreters import xla
 from .lib import xla_client
 from .lib import xla_bridge
 
-def to_dlpack(x):
+def to_dlpack(x: xla.DeviceArray):
   """Returns a DLPack tensor that encapsulates a DeviceArray `x`.
 
   The DLPack shares memory with `x`.
@@ -44,5 +45,7 @@ def from_dlpack(dlpack, backend=None):
   # would be able to figure it out from the DLPack.
   backend = backend or xla_bridge.get_backend()
   buf = xla_client._xla.DLPackManagedTensorToBuffer(dlpack, backend.client)
-  aval = xla._aval_from_xla_shape(buf.shape())
-  return xla.DeviceArray(aval, buf.device(), lazy.array(aval.shape), buf)
+  xla_shape = buf.shape()
+  assert not xla_shape.is_tuple()
+  aval = core.ShapedArray(xla_shape.dimensions(), xla_shape.numpy_dtype())
+  return xla.DeviceArray(aval, buf.device(), lazy.array(aval.shape), buf)  # pytype: disable=attribute-error
