@@ -23,20 +23,36 @@ Adjoint algorithm based on Appendix C of https://arxiv.org/pdf/1806.07366.pdf
 """
 
 
-import functools
+from functools import partial
+import operator as op
 import time
 
 import jax
-from jax.flatten_util import ravel_pytree
-import jax.lax
 import jax.numpy as np
-import jax.ops
-from jax.test_util import check_vjp
+from jax import lax
+from jax import ops
+from jax.util import safe_map, safe_zip
+from jax.flatten_util import ravel_pytree
+from jax.test_util import check_grads
+from jax.tree_util import tree_map
+from jax import linear_util as lu
 import numpy as onp
 import scipy.integrate as osp_integrate
 
+map = safe_map
+zip = safe_zip
 
-@jax.jit
+
+def ravel_first_arg(f, unravel):
+  return ravel_first_arg_(lu.wrap_init(f), unravel).call_wrapped
+
+@lu.transformation
+def ravel_first_arg_(unravel, y_flat, *args):
+  y = unravel(y_flat)
+  ans = yield (y,) + args, {}
+  ans_flat, _ = ravel_pytree(ans)
+  yield ans_flat
+
 def interp_fit_dopri(y0, y1, k, dt):
   # Fit a polynomial to the results of a Runge-Kutta step.
   dps_c_mid = np.array([
@@ -46,562 +62,234 @@ def interp_fit_dopri(y0, y1, k, dt):
   y_mid = y0 + dt * np.dot(dps_c_mid, k)
   return np.array(fit_4th_order_polynomial(y0, y1, y_mid, k[0], k[-1], dt))
 
-
-@jax.jit
 def fit_4th_order_polynomial(y0, y1, y_mid, dy0, dy1, dt):
-  """Fit fourth order polynomial over function interval.
-
-  Args:
-      y0: function value at the start of the interval.
-      y1: function value at the end of the interval.
-      y_mid: function value at the mid-point of the interval.
-      dy0: derivative value at the start of the interval.
-      dy1: derivative value at the end of the interval.
-      dt: width of the interval.
-  Returns:
-      Coefficients `[a, b, c, d, e]` for the polynomial
-      p = a * x ** 4 + b * x ** 3 + c * x ** 2 + d * x + e
-  """
-  v = np.stack([dy0, dy1, y0, y1, y_mid])
-  a = np.dot(np.hstack([-2. * dt, 2. * dt, np.array([-8., -8., 16.])]), v)
-  b = np.dot(np.hstack([5. * dt, -3. * dt, np.array([18., 14., -32.])]), v)
-  c = np.dot(np.hstack([-4. * dt, dt, np.array([-11., -5., 16.])]), v)
+  a = -2.*dt*dy0 + 2.*dt*dy1 -  8.*y0 -  8.*y1 + 16.*y_mid
+  b =  5.*dt*dy0 - 3.*dt*dy1 + 18.*y0 + 14.*y1 - 32.*y_mid
+  c = -4.*dt*dy0 +    dt*dy1 - 11.*y0 -  5.*y1 + 16.*y_mid
   d = dt * dy0
   e = y0
   return a, b, c, d, e
 
-
-@functools.partial(jax.jit, static_argnums=(0,))
 def initial_step_size(fun, t0, y0, order, rtol, atol, f0):
-  """Empirically choose initial step size.
-
-  Args:
-    fun: Function to evaluate like `func(y, t)` to compute the time
-      derivative of `y`.
-    t0: initial time.
-    y0: initial value for the state.
-    order: order of interpolation
-    rtol: relative local error tolerance for solver.
-    atol: absolute local error tolerance for solver.
-    f0: initial value for the derivative, computed from `func(t0, y0)`.
-  Returns:
-    Initial step size for odeint algorithm.
-
-  Algorithm from:
-  E. Hairer, S. P. Norsett G. Wanner,
-  Solving Ordinary Differential Equations I: Nonstiff Problems, Sec. II.4.
-  """
+  # Algorithm from:
+  # E. Hairer, S. P. Norsett G. Wanner,
+  # Solving Ordinary Differential Equations I: Nonstiff Problems, Sec. II.4.
   scale = atol + np.abs(y0) * rtol
   d0 = np.linalg.norm(y0 / scale)
   d1 = np.linalg.norm(f0 / scale)
-  order_pow = (1. / (order + 1.))
 
-  h0 = np.where(np.any(np.asarray([d0 < 1e-5, d1 < 1e-5])),
-                1e-6,
-                0.01 * d0 / d1)
+  h0 = np.where((d0 < 1e-5) | (d1 < 1e-5), 1e-6, 0.01 * d0 / d1)
 
   y1 = y0 + h0 * f0
   f1 = fun(y1, t0 + h0)
   d2 = np.linalg.norm((f1 - f0) / scale) / h0
 
-  h1 = np.where(np.all(np.asarray([d1 <= 1e-15, d2 <= 1e-15])),
+  h1 = np.where((d1 <= 1e-15) & (d2 <= 1e-15),
                 np.maximum(1e-6, h0 * 1e-3),
-                (0.01 / np.max(d1 + d2))**order_pow)
+                (0.01 / np.max(d1 + d2)) ** (1. / (order + 1.)))
 
   return np.minimum(100. * h0, h1)
 
-
-@functools.partial(jax.jit, static_argnums=(0,))
 def runge_kutta_step(func, y0, f0, t0, dt):
-  """Take an arbitrary Runge-Kutta step and estimate error.
-
-  Args:
-      func: Function to evaluate like `func(y, t)` to compute the time
-        derivative of `y`.
-      y0: initial value for the state.
-      f0: initial value for the derivative, computed from `func(t0, y0)`.
-      t0: initial time.
-      dt: time step.
-      alpha, beta, c: Butcher tableau describing how to take the Runge-Kutta
-        step.
-
-  Returns:
-      y1: estimated function at t1 = t0 + dt
-      f1: derivative of the state at t1
-      y1_error: estimated error at t1
-      k: list of Runge-Kutta coefficients `k` used for calculating these terms.
-  """
   # Dopri5 Butcher tableaux
   alpha = np.array([1 / 5, 3 / 10, 4 / 5, 8 / 9, 1., 1., 0])
-  beta = np.array(
-      [[1 / 5, 0, 0, 0, 0, 0, 0],
-       [3 / 40, 9 / 40, 0, 0, 0, 0, 0],
-       [44 / 45, -56 / 15, 32 / 9, 0, 0, 0, 0],
-       [19372 / 6561, -25360 / 2187, 64448 / 6561, -212 / 729, 0, 0, 0],
-       [9017 / 3168, -355 / 33, 46732 / 5247, 49 / 176, -5103 / 18656, 0, 0],
-       [35 / 384, 0, 500 / 1113, 125 / 192, -2187 / 6784, 11 / 84, 0]])
-  c_sol = np.array([35 / 384, 0, 500 / 1113, 125 / 192, -2187 / 6784, 11 / 84,
-                    0])
+  beta = np.array([
+      [1 / 5, 0, 0, 0, 0, 0, 0],
+      [3 / 40, 9 / 40, 0, 0, 0, 0, 0],
+      [44 / 45, -56 / 15, 32 / 9, 0, 0, 0, 0],
+      [19372 / 6561, -25360 / 2187, 64448 / 6561, -212 / 729, 0, 0, 0],
+      [9017 / 3168, -355 / 33, 46732 / 5247, 49 / 176, -5103 / 18656, 0, 0],
+      [35 / 384, 0, 500 / 1113, 125 / 192, -2187 / 6784, 11 / 84, 0]
+  ])
+  c_sol = np.array([35 / 384, 0, 500 / 1113, 125 / 192, -2187 / 6784, 11 / 84, 0])
   c_error = np.array([35 / 384 - 1951 / 21600, 0, 500 / 1113 - 22642 / 50085,
                       125 / 192 - 451 / 720, -2187 / 6784 - -12231 / 42400,
                       11 / 84 - 649 / 6300, -1. / 60.])
 
-  def _fori_body_fun(i, val):
+  def body_fun(i, k):
     ti = t0 + dt * alpha[i-1]
-    yi = y0 + dt * np.dot(beta[i-1, :], val)
+    yi = y0 + dt * np.dot(beta[i-1, :], k)
     ft = func(yi, ti)
-    return jax.ops.index_update(val, jax.ops.index[i, :], ft)
+    return ops.index_update(k, jax.ops.index[i, :], ft)
 
-  k = jax.lax.fori_loop(
-      1,
-      7,
-      _fori_body_fun,
-      jax.ops.index_update(np.zeros((7, f0.shape[0])), jax.ops.index[0, :], f0))
+  k = ops.index_update(np.zeros((7, f0.shape[0])), ops.index[0, :], f0)
+  k = lax.fori_loop(1, 7, body_fun, k)
 
   y1 = dt * np.dot(c_sol, k) + y0
   y1_error = dt * np.dot(c_error, k)
   f1 = k[-1]
   return y1, f1, y1_error, k
 
-
-@jax.jit
 def error_ratio(error_estimate, rtol, atol, y0, y1):
   err_tol = atol + rtol * np.maximum(np.abs(y0), np.abs(y1))
   err_ratio = error_estimate / err_tol
-  return np.mean(err_ratio**2)
+  return np.mean(err_ratio ** 2)
 
-
-@jax.jit
-def optimal_step_size(last_step,
-                      mean_error_ratio,
-                      safety=0.9,
-                      ifactor=10.0,
-                      dfactor=0.2,
-                      order=5.0):
+def optimal_step_size(last_step, mean_error_ratio, safety=0.9, ifactor=10.0,
+                      dfactor=0.2, order=5.0):
   """Compute optimal Runge-Kutta stepsize."""
   mean_error_ratio = np.max(mean_error_ratio)
-  dfactor = np.where(mean_error_ratio < 1,
-                     1.0,
-                     dfactor)
+  dfactor = np.where(mean_error_ratio < 1, 1.0, dfactor)
 
   err_ratio = np.sqrt(mean_error_ratio)
   factor = np.maximum(1.0 / ifactor,
-                      np.minimum(err_ratio**(1.0 / order) / safety,
-                                 1.0 / dfactor))
-  return np.where(mean_error_ratio == 0,
-                  last_step * ifactor,
-                  last_step / factor,)
+                      np.minimum(err_ratio**(1.0 / order) / safety, 1.0 / dfactor))
+  return np.where(mean_error_ratio == 0, last_step * ifactor, last_step / factor)
 
-
-@functools.partial(jax.jit, static_argnums=(0,))
-def odeint(ofunc, y0, t, *args, **kwargs):
+def odeint(func, y0, t, *args, rtol=1.4e-8, atol=1.4e-8, mxstep=np.inf):
   """Adaptive stepsize (Dormand-Prince) Runge-Kutta odeint implementation.
 
   Args:
-    ofunc: Function to evaluate `yt = ofunc(y, t, *args)` that
-      returns the time derivative of `y`.
-    y0: initial value for the state.
-    t: Timespan for `ofunc` evaluation like `np.linspace(0., 10., 101)`.
-    *args: Additional arguments to `ofunc` beyond y0 and t.
-    **kwargs: Two relevant keyword arguments:
-      'rtol': Relative local error tolerance for solver.
-      'atol': Absolute local error tolerance for solver.
-      'mxstep': Maximum number of steps to take for each timepoint.
+    func: function to evaluate the time derivative of the solution `y` at time
+      `t` as `func(y, t, *args)`, producing the same shape/structure as `y0`.
+    y0: array or pytree of arrays representing the initial value for the state.
+    t: array of float times for evaluation, like `np.linspace(0., 10., 101)`,
+      in which the values must be strictly increasing.
+    *args: tuple of additional arguments for `func`.
+    rtol: float, relative local error tolerance for solver (optional).
+    atol: float, absolute local error tolerance for solver (optional).
+    mxstep: int, maximum number of steps to take for each timepoint (optional).
 
   Returns:
-    Integrated system values at each timepoint.
+    Values of the solution `y` (i.e. integrated system values) at each time
+    point in `t`, represented as an array (or pytree of arrays) with the same
+    shape/structure as `y0` except with a new leading axis of length `len(t)`.
   """
-  rtol = kwargs.get('rtol', 1.4e-8)
-  atol = kwargs.get('atol', 1.4e-8)
-  mxstep = kwargs.get('mxstep', np.inf)
+  return _odeint_wrapper(func, rtol, atol, mxstep, y0, t, *args)
 
-  @functools.partial(jax.jit, static_argnums=(0,))
-  def _fori_body_fun(func, i, val):
-    """Internal fori_loop body to interpolate an integral at each timestep."""
-    t, cur_y, cur_f, cur_t, dt, last_t, interp_coeff, solution = val
-    cur_y, cur_f, cur_t, dt, last_t, interp_coeff, _ = jax.lax.while_loop(
-        lambda x: (x[2] < t[i]) & (x[-1] < mxstep),
-        functools.partial(_while_body_fun, func),
-        (cur_y, cur_f, cur_t, dt, last_t, interp_coeff, 0.))
+@partial(jax.jit, static_argnums=(0, 1, 2, 3))
+def _odeint_wrapper(func, rtol, atol, mxstep, y0, ts, *args):
+  y0, unravel = ravel_pytree(y0)
+  func = ravel_first_arg(func, unravel)
+  out = _odeint(func, rtol, atol, mxstep, y0, ts, *args)
+  return jax.vmap(unravel)(out)
 
-    relative_output_time = (t[i] - last_t) / (cur_t - last_t)
-    out_x = np.polyval(interp_coeff, relative_output_time)
+@partial(jax.custom_vjp, nondiff_argnums=(0, 1, 2, 3))
+def _odeint(func, rtol, atol, mxstep, y0, ts, *args):
+  func_ = lambda y, t: func(y, t, *args)
 
-    return (t, cur_y, cur_f, cur_t, dt, last_t, interp_coeff,
-            jax.ops.index_update(solution,
-                                 jax.ops.index[i, :],
-                                 out_x))
+  def scan_fun(carry, target_t):
 
-  @functools.partial(jax.jit, static_argnums=(0,))
-  def _while_body_fun(func, x):
-    """Internal while_loop body to determine interpolation coefficients."""
-    cur_y, cur_f, cur_t, dt, last_t, interp_coeff, j = x
-    next_t = cur_t + dt
-    next_y, next_f, next_y_error, k = runge_kutta_step(
-        func, cur_y, cur_f, cur_t, dt)
-    error_ratios = error_ratio(next_y_error, rtol, atol, cur_y, next_y)
-    new_interp_coeff = interp_fit_dopri(cur_y, next_y, k, dt)
-    dt = optimal_step_size(dt, error_ratios)
+    def cond_fun(state):
+      i, _, _, t, _, _, _ = state
+      return (t < target_t) & (i < mxstep)
 
-    next_j = j + 1
-    new_rav, unravel = ravel_pytree(
-        (next_y, next_f, next_t, dt, cur_t, new_interp_coeff, next_j))
-    old_rav, _ = ravel_pytree(
-        (cur_y, cur_f, cur_t, dt, last_t, interp_coeff, next_j))
+    def body_fun(state):
+      i, y, f, t, dt, last_t, interp_coeff = state
+      next_y, next_f, next_y_error, k = runge_kutta_step(func_, y, f, t, dt)
+      next_t = t + dt
+      error_ratios = error_ratio(next_y_error, rtol, atol, y, next_y)
+      new_interp_coeff = interp_fit_dopri(y, next_y, k, dt)
+      dt = optimal_step_size(dt, error_ratios)
 
-    return unravel(np.where(np.all(error_ratios <= 1.),
-                            new_rav,
-                            old_rav))
+      new = [i + 1, next_y, next_f, next_t, dt,      t, new_interp_coeff]
+      old = [i + 1,      y,      f,      t, dt, last_t,     interp_coeff]
+      return map(partial(np.where, np.all(error_ratios <= 1.)), new, old)
 
-  func = lambda y, t: ofunc(y, t, *args)
-  f0 = func(y0, t[0])
-  dt = initial_step_size(func, t[0], y0, 4, rtol, atol, f0)
+    _, *carry = lax.while_loop(cond_fun, body_fun, [0] + carry)
+    _, _, t, _, last_t, interp_coeff = carry
+    relative_output_time = (target_t - last_t) / (t - last_t)
+    y_target = np.polyval(interp_coeff, relative_output_time)
+    return carry, y_target
+
+  f0 = func_(y0, ts[0])
+  dt = initial_step_size(func_, ts[0], y0, 4, rtol, atol, f0)
   interp_coeff = np.array([y0] * 5)
+  init_carry = [y0, f0, ts[0], dt, ts[0], interp_coeff]
+  _, ys = lax.scan(scan_fun, init_carry, ts[1:])
+  return np.concatenate((y0[None], ys))
 
-  return jax.lax.fori_loop(1,
-                           t.shape[0],
-                           functools.partial(_fori_body_fun, func),
-                           (t, y0, f0, t[0], dt, t[0], interp_coeff,
-                            jax.ops.index_update(
-                                np.zeros((t.shape[0], y0.shape[0])),
-                                jax.ops.index[0, :],
-                                y0)))[-1]
+def _odeint_fwd(func, rtol, atol, mxstep, y0, ts, *args):
+  ys = _odeint(func, rtol, atol, mxstep, y0, ts, *args)
+  return ys, (ys, ts, args)
 
+def _odeint_rev(func, rtol, atol, mxstep, res, g):
+  ys, ts, args = res
 
-def vjp_odeint(ofunc, y0, t, *args, **kwargs):
-  """Return a function that calculates `vjp(odeint(func(y, t, *args))`.
-
-  Args:
-    ofunc: Function `ydot = ofunc(y, t, *args)` to compute the time
-      derivative of `y`.
-    y0: initial value for the state.
-    t: Timespan for `ofunc` evaluation like `np.linspace(0., 10., 101)`.
-    *args: Additional arguments to `ofunc` beyond y0 and t.
-    **kwargs: Two relevant keyword arguments:
-      'rtol': Relative local error tolerance for solver.
-      'atol': Absolute local error tolerance for solver.
-      'mxstep': Maximum number of steps to take for each timepoint.
-
-  Returns:
-    VJP function `vjp = vjp_all(g)` where `yt = ofunc(y, t, *args)`
-    and g is used for VJP calculation. To evaluate the gradient w/ the VJP,
-    supply `g = np.ones_like(yt)`. To evaluate the reverse Jacobian do a vmap
-    over the standard basis of yt.
-  """
-  rtol = kwargs.get('rtol', 1.4e-8)
-  atol = kwargs.get('atol', 1.4e-8)
-  mxstep = kwargs.get('mxstep', np.inf)
-
-  flat_args, unravel_args = ravel_pytree(args)
-  flat_func = lambda y, t, flat_args: ofunc(y, t, *unravel_args(flat_args))
-
-  @jax.jit
-  def aug_dynamics(augmented_state, t, flat_args):
+  def aug_dynamics(augmented_state, t, *args):
     """Original system augmented with vjp_y, vjp_t and vjp_args."""
-    state_len = int(np.floor_divide(
-        augmented_state.shape[0] - flat_args.shape[0] - 1, 2))
-    y = augmented_state[:state_len]
-    adjoint = augmented_state[state_len:2*state_len]
-    dy_dt, vjpfun = jax.vjp(flat_func, y, t, flat_args)
-    return np.hstack([np.ravel(dy_dt), np.hstack(vjpfun(-adjoint))])
+    y, y_bar, *_ = augmented_state
+    y_dot, vjpfun = jax.vjp(func, y, -t, *args)
+    return (-y_dot, *vjpfun(y_bar))
 
-  rev_aug_dynamics = lambda y, t, flat_args: -aug_dynamics(y, -t, flat_args)
+  y_bar = g[-1]
+  ts_bar = []
+  t0_bar = 0.
 
-  @jax.jit
-  def _fori_body_fun(i, val):
-    """fori_loop function for VJP calculation."""
-    rev_yt, rev_t, rev_tarray, rev_gi, vjp_y, vjp_t0, vjp_args, time_vjp_list = val
-    this_yt = rev_yt[i, :]
-    this_t = rev_t[i]
-    this_tarray = rev_tarray[i, :]
-    this_gi = rev_gi[i, :]
-    # this is g[i-1, :] when g has been reversed
-    this_gim1 = rev_gi[i+1, :]
-    state_len = this_yt.shape[0]
-    vjp_cur_t = np.dot(flat_func(this_yt, this_t, flat_args), this_gi)
-    vjp_t0 = vjp_t0 - vjp_cur_t
-    # Run augmented system backwards to the previous observation.
-    aug_y0 = np.hstack((this_yt, vjp_y, vjp_t0, vjp_args))
-    aug_ans = odeint(rev_aug_dynamics,
-                     aug_y0,
-                     this_tarray,
-                     flat_args,
-                     rtol=rtol,
-                     atol=atol,
-                     mxstep=mxstep)
-    vjp_y = aug_ans[1][state_len:2*state_len] + this_gim1
-    vjp_t0 = aug_ans[1][2*state_len]
-    vjp_args = aug_ans[1][2*state_len+1:]
-    time_vjp_list = jax.ops.index_update(time_vjp_list, i, vjp_cur_t)
-    return rev_yt, rev_t, rev_tarray, rev_gi, vjp_y, vjp_t0, vjp_args, time_vjp_list
+  def scan_fun(carry, i):
+    y_bar, t0_bar, args_bar = carry
+    # Compute effect of moving measurement time
+    t_bar = np.dot(func(ys[i], ts[i], *args), g[i])
+    t0_bar = t0_bar - t_bar
+    # Run augmented system backwards to previous observation
+    _, y_bar, t0_bar, args_bar = odeint(
+        aug_dynamics, (ys[i], y_bar, t0_bar, args_bar), np.array([ts[i - 1], ts[i]]),
+        *args, rtol=rtol, atol=atol, mxstep=mxstep)
+    y_bar, t0_bar, args_bar = tree_map(op.itemgetter(1), (y_bar, t0_bar, args_bar))
+    # Add gradient from current output
+    y_bar = y_bar + g[i - 1]
+    return (y_bar, t0_bar, args_bar), t_bar
 
-  @jax.jit
-  def vjp_all(g, yt, t):
-    """Calculate the VJP g * Jac(odeint(ofunc, y0, t, *args))."""
-    rev_yt = yt[-1::-1, :]
-    rev_t = t[-1::-1]
-    rev_tarray = -np.array([t[-1:0:-1], t[-2::-1]]).T
-    rev_gi = g[-1::-1, :]
+  init_carry = (g[-1], 0., tree_map(np.zeros_like, args))
+  (y_bar, t0_bar, args_bar), rev_ts_bar = lax.scan(
+      scan_fun, init_carry, np.arange(len(ts) - 1, 0, -1))
+  ts_bar = np.concatenate([np.array([t0_bar]), rev_ts_bar[::-1]])
+  return (y_bar, ts_bar, *args_bar)
 
-    vjp_y = g[-1, :]
-    vjp_t0 = 0.
-    vjp_args = np.zeros_like(flat_args)
-    time_vjp_list = np.zeros_like(t)
-
-    result = jax.lax.fori_loop(0,
-                               rev_t.shape[0]-1,
-                               _fori_body_fun,
-                               (rev_yt,
-                                rev_t,
-                                rev_tarray,
-                                rev_gi,
-                                vjp_y,
-                                vjp_t0,
-                                vjp_args,
-                                time_vjp_list))
-
-    time_vjp_list = jax.ops.index_update(result[-1], -1, result[-3])
-    vjp_times = np.hstack(time_vjp_list)[::-1]
-
-    return tuple([result[-4], vjp_times] + list(result[-2]))
-
-  primals_out = odeint(flat_func, y0, t, flat_args, rtol=rtol, atol=atol, mxstep=mxstep)
-  vjp_fun = lambda g: vjp_all(g, primals_out, t)
-
-  return primals_out, vjp_fun
+_odeint.defvjp(_odeint_fwd, _odeint_rev)
 
 
-def build_odeint(ofunc, rtol=1.4e-8, atol=1.4e-8, mxstep=onp.inf):
-  """Return `f(y0, t, args) = odeint(ofunc(y, t, *args), y0, t, args)`.
-
-  Given the function ofunc(y, t, *args), return the jitted function
-  `f(y0, t, args) = odeint(ofunc(y, t, *args), y0, t, args)` with
-  the VJP of `f` defined using `vjp_odeint`, where:
-
-    `y0` is the initial condition of the ODE integration,
-    `t` is the time course of the integration, and
-    `*args` are all other arguments to `ofunc`.
-
-  Args:
-    ofunc: The function to be wrapped into an ODE integration.
-    rtol: relative local error tolerance for solver.
-    atol: absolute local error tolerance for solver.
-    mxstep: Maximum number of steps to take for each timepoint.
-
-  Returns:
-    `f(y0, t, args) = odeint(ofunc(y, t, *args), y0, t, args)`
-  """
-  ct_odeint = jax.custom_transforms(
-      lambda y0, t, *args: odeint(ofunc, y0, t, *args, rtol=rtol, atol=atol, mxstep=mxstep))
-
-  v = lambda y0, t, *args: vjp_odeint(ofunc, y0, t, *args, rtol=rtol, atol=atol, mxstep=mxstep)
-  jax.defvjp_all(ct_odeint, v)
-
-  return jax.jit(ct_odeint)
-
-
-def my_odeint_grad(fun):
-  """Calculate the Jacobian of an odeint."""
-  @jax.jit
-  def _gradfun(*args, **kwargs):
-    ys, pullback = vjp_odeint(fun, *args, **kwargs)
-    my_grad = pullback(np.ones_like(ys))
-    return my_grad
-  return _gradfun
-
-
-def my_odeint_jacrev(fun):
-  """Calculate the Jacobian of an odeint."""
-  @jax.jit
-  def _jacfun(*args, **kwargs):
-    ys, pullback = vjp_odeint(fun, *args, **kwargs)
-    my_jac = jax.vmap(pullback)(jax.api._std_basis(ys))
-    my_jac = jax.api.tree_map(
-        functools.partial(jax.api._unravel_array_into_pytree, ys, 0), my_jac)
-    my_jac = jax.api.tree_transpose(
-        jax.api.tree_structure(args), jax.api.tree_structure(ys), my_jac)
-    return my_jac
-  return _jacfun
-
-
-def nd(f, x, eps=0.0001):
-  flat_x, unravel = ravel_pytree(x)
-  dim = len(flat_x)
-  g = onp.zeros_like(flat_x)
-  for i in range(dim):
-    d = onp.zeros_like(flat_x)
-    d[i] = eps
-    g[i] = (f(unravel(flat_x + d)) - f(unravel(flat_x - d))) / (2.0 * eps)
-  return g
-
-
-def test_grad_vjp_odeint():
-  """Compare numerical and exact differentiation of a simple odeint."""
-
-  def f(y, t, arg1, arg2):
-    return -np.sqrt(t) - y + arg1 - np.mean((y + arg2)**2)
-
-  def onearg_odeint(args):
-    return np.sum(
-        odeint(f, *args, atol=1e-8, rtol=1e-8))
-
-  dim = 10
-  t0 = 0.1
-  t1 = 0.2
-  y0 = np.linspace(0.1, 0.9, dim)
-  arg1 = 0.1
-  arg2 = 0.2
-  wrap_args = (y0, np.array([t0, t1]), arg1, arg2)
-
-  numerical_grad = nd(onearg_odeint, wrap_args)
-  exact_grad, _ = ravel_pytree(my_odeint_grad(f)(*wrap_args))
-
-  assert np.allclose(numerical_grad, exact_grad)
-
-
-def plot_gradient_field(ax, func, xlimits, ylimits, numticks=30):
-  """Plot the gradient field of `func` on `ax`."""
-  x = np.linspace(*xlimits, num=numticks)
-  y = np.linspace(*ylimits, num=numticks)
-  x_mesh, y_mesh = np.meshgrid(x, y)
-  zs = jax.vmap(func)(y_mesh.ravel(), x_mesh.ravel())
-  z_mesh = zs.reshape(x_mesh.shape)
-  ax.quiver(x_mesh, y_mesh, np.ones(z_mesh.shape), z_mesh)
-  ax.set_xlim(xlimits)
-  ax.set_ylim(ylimits)
-
-
-@jax.jit
-def pend(y, t, arg1, arg2):
-  """Simple pendulum system for odeint testing."""
-  del t
+def pend(np, y, _, m, g):
   theta, omega = y
-  dydt = np.array([omega, -arg1*omega - arg2*np.sin(theta)])
-  return dydt
-
-
-@jax.jit
-def swoop(y, t, arg1, arg2):
-  return np.array(y - np.sin(t) - np.cos(t) * arg1 + arg2)
-
-
-@jax.jit
-def decay(y, t, arg1, arg2):
-  return -np.sqrt(t) - y + arg1 - np.mean((y + arg2)**2)
-
+  return [omega, -m * omega - g * np.sin(theta)]
 
 def benchmark_odeint(fun, y0, tspace, *args):
   """Time performance of JAX odeint method against scipy.integrate.odeint."""
-  n_trials = 5
+  n_trials = 10
+  n_repeat = 100
+  y0, tspace = onp.array(y0), onp.array(tspace)
+  onp_fun = partial(fun, onp)
+  scipy_times = []
   for k in range(n_trials):
     start = time.time()
-    scipy_result = osp_integrate.odeint(fun, y0, tspace, args)
+    for _ in range(n_repeat):
+      scipy_result = osp_integrate.odeint(onp_fun, y0, tspace, args)
     end = time.time()
-    print('scipy odeint elapsed time ({} of {}): {}'.format(
-        k+1, n_trials, end-start))
+    print('scipy odeint elapsed time ({} of {}): {}'.format(k+1, n_trials, end-start))
+    scipy_times.append(end - start)
+  y0, tspace = np.array(y0), np.array(tspace)
+  jax_fun = partial(fun, np)
+  jax_times = []
   for k in range(n_trials):
     start = time.time()
-    jax_result = odeint(fun, np.array(y0), np.array(tspace), *args)
+    for _ in range(n_repeat):
+      jax_result = odeint(jax_fun, y0, tspace, *args)
     jax_result.block_until_ready()
     end = time.time()
-    print('JAX odeint elapsed time ({} of {}): {}'.format(
-        k+1, n_trials, end-start))
+    print('JAX odeint elapsed time ({} of {}): {}'.format(k+1, n_trials, end-start))
+    jax_times.append(end - start)
+  print('(avg scipy time) / (avg jax time) = {}'.format(
+      onp.mean(scipy_times[1:]) / onp.mean(jax_times[1:])))
   print('norm(scipy result-jax result): {}'.format(
       np.linalg.norm(np.asarray(scipy_result) - jax_result)))
-
   return scipy_result, jax_result
 
-
 def pend_benchmark_odeint():
-  _, _ = benchmark_odeint(pend,
-                          (onp.pi - 0.1, 0.0),
-                          onp.linspace(0., 10., 101),
-                          0.25,
-                          9.8)
+  _, _ = benchmark_odeint(pend, [np.pi - 0.1, 0.0], np.linspace(0., 10., 101),
+                          0.25, 9.8)
 
+def pend_check_grads():
+  def f(y0, ts, *args):
+    return odeint(partial(pend, np), y0, ts, *args)
 
-def test_odeint_grad():
-  """Test the gradient behavior of various ODE integrations."""
-  def _test_odeint_grad(func, *args):
-    def onearg_odeint(fargs):
-      return np.sum(odeint(func, *fargs))
+  y0 = [np.pi - 0.1, 0.0]
+  ts = np.linspace(0., 1., 11)
+  args = (0.25, 9.8)
 
-    numerical_grad = nd(onearg_odeint, args)
-    exact_grad, _ = ravel_pytree(my_odeint_grad(func)(*args))
-    assert np.allclose(numerical_grad, exact_grad)
-
-  ts = np.array((0.1, 0.2))
-  y0 = np.linspace(0.1, 0.9, 10)
-  big_y0 = np.linspace(1.1, 10.9, 10)
-
-  # check pend()
-  for cond in (
-      (np.array((onp.pi - 0.1, 0.0)), ts, 0.25, 0.98),
-      (np.array((onp.pi * 0.1, 0.0)), ts, 0.1, 0.4),
-      ):
-    _test_odeint_grad(pend, *cond)
-
-  # check swoop
-  for cond in (
-      (y0, ts, 0.1, 0.2),
-      (big_y0, ts, 0.1, 0.3),
-      ):
-    _test_odeint_grad(swoop, *cond)
-
-  # check decay
-  for cond in (
-      (y0, ts, 0.1, 0.2),
-      (big_y0, ts, 0.1, 0.3),
-      ):
-    _test_odeint_grad(decay, *cond)
-
-
-def test_odeint_vjp():
-  """Use check_vjp to check odeint VJP calculations."""
-
-  # check pend()
-  y = np.array([np.pi - 0.1, 0.0])
-  t = np.linspace(0., 10., 11)
-  b = 0.25
-  c = 9.8
-  wrap_args = (y, t, b, c)
-  pend_odeint_wrap = lambda y, t, *args: odeint(pend, y, t, *args)
-  pend_vjp_wrap = lambda y, t, *args: vjp_odeint(pend, y, t, *args)
-  check_vjp(pend_odeint_wrap, pend_vjp_wrap, wrap_args)
-
-  # check swoop()
-  y = np.array([0.1])
-  t = np.linspace(0., 10., 11)
-  arg1 = 0.1
-  arg2 = 0.2
-  wrap_args = (y, t, arg1, arg2)
-  swoop_odeint_wrap = lambda y, t, *args: odeint(swoop, y, t, *args)
-  swoop_vjp_wrap = lambda y, t, *args: vjp_odeint(swoop, y, t, *args)
-  check_vjp(swoop_odeint_wrap, swoop_vjp_wrap, wrap_args)
-
-  # decay() check_vjp hangs!
-
-
-def test_defvjp_all():
-  """Use build_odeint to check odeint VJP calculations."""
-  n_trials = 5
-  swoop_build = build_odeint(swoop)
-  jacswoop = jax.jit(jax.jacrev(swoop_build))
-  y = np.array([0.1])
-  t = np.linspace(0., 2., 11)
-  arg1 = 0.1
-  arg2 = 0.2
-  wrap_args = (y, t, arg1, arg2)
-  for k in range(n_trials):
-    start = time.time()
-    rslt = jacswoop(*wrap_args)
-    rslt.block_until_ready()
-    end = time.time()
-    print('JAX jacrev elapsed time ({} of {}): {}'.format(
-        k+1, n_trials, end-start))
+  check_grads(f, (y0, ts, *args), modes=["rev"], order=2,
+              atol=1e-1, rtol=1e-1)
 
 
 if __name__ == '__main__':
-
-  test_odeint_grad()
-  test_odeint_vjp()
+  pend_benchmark_odeint()
+  pend_check_grads()
