@@ -13,12 +13,13 @@
 # limitations under the License.
 
 from functools import partial
+from typing import Callable, Dict, Type
 
 from absl import logging
 import numpy as onp
 
 from .. import core
-from ..abstract_arrays import ShapedArray, ConcreteArray, array_types, abstract_token
+from ..abstract_arrays import ShapedArray, ConcreteArray, array_types
 from . import partial_eval as pe
 from . import xla
 from .. import linear_util as lu
@@ -86,7 +87,7 @@ def _pvals_to_results_handler(nrep, npar, partitions, out_pvals):
     buffers = [[[None] * npar for _ in range(nrep)] for _ in range(nouts)]
     for raw_idx, tuple_buf in enumerate(out_bufs):
       r, p = onp.unravel_index(raw_idx, (nrep, npar))
-      for i, buf in enumerate(tuple_buf.destructure()):
+      for i, buf in enumerate(tuple_buf):
         buffers[i][r][p] = buf
     return [h(bufs) for h, bufs in zip(handlers, buffers)]
 
@@ -105,7 +106,7 @@ def _aval_to_result_handler(partition, aval):
   return result_handlers[type(aval)](partition, aval)
 
 
-result_handlers = {}
+result_handlers: Dict[Type[core.AbstractValue], Callable] = {}
 
 
 def _array_result_handler(partition, aval):
@@ -134,16 +135,11 @@ result_handlers[ConcreteArray] = _array_result_handler
 
 
 @lu.cache
-def _sharded_callable(fun, partitions, name, *abstract_args):
+def _sharded_callable(fun: lu.WrappedFun, partitions, name, *abstract_args):
   nrep = 1  # TODO generalize
 
   in_pvals = [pe.PartialVal((aval, core.unit)) for aval in abstract_args]
-  with core.new_master(pe.JaxprTrace, True) as master:
-    jaxpr, (out_pvals, consts,
-            env) = pe.trace_to_subjaxpr(fun, master,
-                                        False).call_wrapped(in_pvals)
-    assert not env  # no subtraces here
-    del master, env
+  jaxpr, out_pvals, consts = pe.trace_to_jaxpr(fun, in_pvals, instantiate=False, bottom=True)
 
   if not jaxpr.eqns and all(outvar is core.unitvar for outvar in jaxpr.outvars):
     return lambda *_: [core.unit] * len(jaxpr.outvars)
@@ -183,8 +179,9 @@ def _sharded_callable(fun, partitions, name, *abstract_args):
                  handle_outs)
 
 
-def _sharded_jit_translation_rule(c, jaxpr, axis_env, freevar_nodes,
-                                  in_nodes, name_stack, partitions, backend, name):
+def _sharded_jit_translation_rule(c, axis_env, freevar_nodes,
+                                  in_nodes, name_stack, partitions, backend,
+                                  name, call_jaxpr):
   subc = xb.make_computation_builder("jaxpr_subcomputation")  # TODO(mattjj): name
   freevars = [subc.ParameterWithShape(c.GetShape(n)) for n in freevar_nodes]
 
@@ -195,7 +192,7 @@ def _sharded_jit_translation_rule(c, jaxpr, axis_env, freevar_nodes,
     subc._builder.ClearSharding()
   # args = [subc.ParameterWithShape(c.GetShape(n)) for n in in_nodes]
 
-  out_nodes = xla.jaxpr_subcomp(subc, jaxpr, backend, axis_env, (), freevars, name_stack, *args)
+  out_nodes = xla.jaxpr_subcomp(subc, call_jaxpr, backend, axis_env, (), freevars, name_stack, *args)
 
   subc._builder.SetSharding(_sharding_to_proto(partitions[1]))
   out_tuple = subc.Tuple(*out_nodes)
@@ -206,7 +203,8 @@ def _sharded_jit_translation_rule(c, jaxpr, axis_env, freevar_nodes,
 
 def _execute_spatially_partitioned(compiled, in_handler, out_handler, *args):
   input_bufs = in_handler(args)
-  out_bufs = compiled.ExecuteOnLocalDevices(list(input_bufs))
+  out_bufs = compiled.ExecuteOnLocalDevices(
+      list(input_bufs), tuple_arguments=False)
   return out_handler(out_bufs)
 
 
@@ -267,7 +265,7 @@ def jaxpr_partitions(jaxpr):
 ### sharded_call
 
 
-def _sharded_call_impl(fun, *args, **params):
+def _sharded_call_impl(fun: lu.WrappedFun, *args, **params):
   partitions = params.pop("partitions")
   name = params.pop("name")
   assert not params, params
@@ -277,6 +275,7 @@ def _sharded_call_impl(fun, *args, **params):
 
 
 sharded_call_p = core.Primitive("sharded_call")
+sharded_call_p.call_primitive = True
 sharded_call_p.multiple_results = True
 sharded_call = partial(core.call_bind, sharded_call_p)
 sharded_call_p.def_custom_bind(sharded_call)
