@@ -20,6 +20,7 @@ import textwrap
 
 import scipy.ndimage
 
+from .. import api
 from ..numpy import lax_numpy as np
 from ..numpy.lax_numpy import _wraps
 from ..util import safe_zip as zip
@@ -52,17 +53,8 @@ def _linear_indices_and_weights(coordinate):
   return [(l_index, l_weight), (u_index, u_weight)]
 
 
-@_wraps(scipy.ndimage.map_coordinates, lax_description=textwrap.dedent("""\
-    Only linear interpolation (``order=1``) and modes ``'constant'``,
-    ``'nearest'`` and ``'wrap'`` are currently supported. Note that
-    interpolation near boundaries differs from the scipy function, because we
-    fixed an outstanding bug (https://github.com/scipy/scipy/issues/2640);
-    this function interprets the ``mode`` argument as documented by SciPy, but
-    not as implemented by SciPy.
-    """))
-def map_coordinates(
-    input, coordinates, order, mode='constant', cval=0.0,
-):
+@functools.partial(api.jit, static_argnums=(2, 3, 4))
+def _map_coordinates(input, coordinates, order, mode, cval):
   input = np.asarray(input)
   coordinates = [np.asarray(c, input.dtype) for c in coordinates]
   cval = np.asarray(cval, input.dtype)
@@ -111,3 +103,17 @@ def map_coordinates(
     outputs.append(_nonempty_prod(weights) * contribution)
   result = _nonempty_sum(outputs)
   return result
+
+
+@_wraps(scipy.ndimage.map_coordinates, lax_description=textwrap.dedent("""\
+    Only linear interpolation (``order=1``) and modes ``'constant'``,
+    ``'nearest'`` and ``'wrap'`` are currently supported. Note that
+    interpolation near boundaries differs from the scipy function, because we
+    fixed an outstanding bug (https://github.com/scipy/scipy/issues/2640);
+    this function interprets the ``mode`` argument as documented by SciPy, but
+    not as implemented by SciPy.
+    """))
+def map_coordinates(
+    input, coordinates, order, mode='constant', cval=0.0,
+):
+  return _map_coordinates(input, coordinates, order, mode, cval)
