@@ -13,30 +13,30 @@
 # limitations under the License.
 
 
-import numpy as onp
+import numpy as np
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy import lax_numpy as np
+from ...numpy import lax_numpy as jnp
 from ..special import gammaln, xlogy
 
 
 def _is_simplex(x):
-    x_sum = np.sum(x, axis=-1)
-    return np.all(x > 0, axis=-1) & (x_sum <= 1) & (x_sum > 1 - 1e-6)
+    x_sum = jnp.sum(x, axis=-1)
+    return jnp.all(x > 0, axis=-1) & (x_sum <= 1) & (x_sum > 1 - 1e-6)
 
 
-@np._wraps(osp_stats.dirichlet.logpdf, update_doc=False)
+@jnp._wraps(osp_stats.dirichlet.logpdf, update_doc=False)
 def logpdf(x, alpha):
-    args = (onp.ones((0,), lax.dtype(x)), onp.ones((1,), lax.dtype(alpha)))
+    args = (np.ones((0,), lax.dtype(x)), np.ones((1,), lax.dtype(alpha)))
     to_dtype = lax.dtype(osp_stats.dirichlet.logpdf(*args))
     x, alpha = [lax.convert_element_type(arg, to_dtype) for arg in (x, alpha)]
-    one = np._constant_like(x, 1)
-    normalize_term = np.sum(gammaln(alpha), axis=-1) - gammaln(np.sum(alpha, axis=-1))
-    log_probs = lax.sub(np.sum(xlogy(lax.sub(alpha, one), x), axis=-1), normalize_term)
-    return np.where(_is_simplex(x), log_probs, -np.inf)
+    one = jnp._constant_like(x, 1)
+    normalize_term = jnp.sum(gammaln(alpha), axis=-1) - gammaln(jnp.sum(alpha, axis=-1))
+    log_probs = lax.sub(jnp.sum(xlogy(lax.sub(alpha, one), x), axis=-1), normalize_term)
+    return jnp.where(_is_simplex(x), log_probs, -jnp.inf)
 
 
-@np._wraps(osp_stats.dirichlet.pdf, update_doc=False)
+@jnp._wraps(osp_stats.dirichlet.pdf, update_doc=False)
 def pdf(x, alpha):
   return lax.exp(logpdf(x, alpha))
