@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-import numpy as onp
+import numpy as np
 import scipy.stats as osp_stats
 
 from ... import lax
@@ -27,7 +27,7 @@ def logpdf(x, df, loc=0, scale=1):
   scaled_x = lax.div(lax.sub(x, loc), scale)
   df_over_two = lax.div(df, two)
   df_plus_one_over_two = lax.add(df_over_two, _constant_like(x, 0.5))
-  normalize_term_const = lax.mul(lax.mul(scale, scale), _constant_like(x, onp.pi))
+  normalize_term_const = lax.mul(lax.mul(scale, scale), _constant_like(x, np.pi))
   normalize_term_tmp = lax.div(lax.log(lax.mul(normalize_term_const, df)), two)
   normalize_term = lax.sub(lax.add(lax.lgamma(df_over_two), normalize_term_tmp),
                            lax.lgamma(df_plus_one_over_two))
