@@ -349,7 +349,9 @@ class LaxControlFlowTest(jtu.JaxTestCase):
     self.assertAllClose(ans, expected, check_dtypes=False)
 
   def testForiLoopBatchedIssue1190(self):
-    f = lambda x: lax.fori_loop(0, 4, lambda _, x: x + 1, x)
+    cond_fun = lambda carry: carry[0] < 4
+    body_fun = lambda carry: (carry[0] + 1, carry[1] + 1)
+    f = lambda x: lax.while_loop(cond_fun, body_fun, (0, x))
     jaxpr = api.make_jaxpr(api.vmap(f))(np.arange(3))
     eqn = jaxpr.jaxpr.eqns[0]
     self.assertIs(eqn.primitive, lax.while_p)
@@ -820,6 +822,19 @@ class LaxControlFlowTest(jtu.JaxTestCase):
     expected = f(4.)
     self.assertAllClose(y, expected, check_dtypes=False)
 
+  def testCondVmapGrad(self):
+    # https://github.com/google/jax/issues/2264
+    def f_1(x): return x ** 2
+    def f_2(x): return x ** 3
+
+    def f(x): return lax.cond(x > 0, x, f_1, x, f_2)
+    def g(x): return np.where(x > 0, f_1(x), f_2(x))
+
+    x = np.linspace(-1, 1, 20)
+    ans = api.vmap(api.grad(f))(x)
+    expected = api.vmap(api.grad(g))(x)
+    self.assertAllClose(ans, expected, check_dtypes=False)
+
   def testIssue1263(self):
     def f(rng, x):
       cond = random.bernoulli(rng)
@@ -1255,6 +1270,12 @@ class LaxControlFlowTest(jtu.JaxTestCase):
     expected = xs ** 2
     actual = lax.map(f, xs)
     self.assertAllClose(actual, expected, check_dtypes=True)
+
+  def testMapEmpty(self):
+    # https://github.com/google/jax/issues/2412
+    ans = lax.map(lambda x: x * x, np.array([]))
+    expected = np.array([])
+    self.assertAllClose(ans, expected, check_dtypes=True)
 
   def testCaching(self):
     def cond(x):
