@@ -128,6 +128,8 @@ class NumpyLinalgTest(jtu.JaxTestCase):
   def testTensorsolve(self, m, nq, dtype, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
+    if m == 23:
+      _skip_on_mac_xla_bug()
     
     # According to numpy docs the shapes are as follows:
     # Coefficient tensor (a), of shape b.shape + Q. 
@@ -431,7 +433,7 @@ class NumpyLinalgTest(jtu.JaxTestCase):
              (isinstance(axis, tuple) and len(axis) == 1)
           else [None, 'fro', 1, 2, -1, -2, np.inf, -np.inf, 'nuc'])
       for dtype in float_types + complex_types
-      for rng_factory in [jtu.rand_default]))
+      for rng_factory in [jtu.rand_default]))  # type: ignore
   def testNorm(self, shape, dtype, ord, axis, keepdims, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
@@ -584,6 +586,8 @@ class NumpyLinalgTest(jtu.JaxTestCase):
       for shape in [(1, 1), (4, 4), (2, 3, 5), (5, 5, 5), (20, 20), (5, 10)]
       for pnorm in [np.inf, -np.inf, 1, -1, 2, -2, 'fro']
       for dtype in float_types + complex_types))
+  @jtu.skip_on_devices("tpu")  # SVD is not implemented on the TPU backend
+  @jtu.skip_on_devices("gpu")  # TODO(#2203): numerical errors
   def testCond(self, shape, pnorm, dtype):
     _skip_if_unsupported_type(dtype)
 
@@ -617,6 +621,8 @@ class NumpyLinalgTest(jtu.JaxTestCase):
       for rng_factory in [jtu.rand_default]))
   def testTensorinv(self, shape, dtype, rng_factory):
     _skip_if_unsupported_type(dtype)
+    if shape[0] > 100:
+      _skip_on_mac_xla_bug()
     rng = rng_factory()
 
     def tensor_maker():
@@ -719,6 +725,7 @@ class NumpyLinalgTest(jtu.JaxTestCase):
       for dtype in float_types + complex_types
       for n in [-5, -2, -1, 0, 1, 2, 3, 4, 5, 10]
       for rng_factory in [jtu.rand_default]))
+  @jtu.skip_on_devices("tpu")  # TODO(b/149870255): Bug in XLA:TPU?.
   def testMatrixPower(self, shape, dtype, n, rng_factory):
     rng = rng_factory()
     _skip_if_unsupported_type(dtype)
@@ -902,7 +909,7 @@ class ScipyLinalgTest(jtu.JaxTestCase):
       for lhs_shape, rhs_shape in [
           ((1, 1), (1, 1)),
           ((4, 4), (4,)),
-          ((8, 8), (8, 4, 2)),
+          ((8, 8), (8, 4)),
       ]
       for trans in [0, 1, 2]
       for dtype in float_types + complex_types
@@ -1030,8 +1037,10 @@ class ScipyLinalgTest(jtu.JaxTestCase):
       for conjugate_a in (
           [False] if np.issubdtype(dtype, np.floating) else [False, True])
       for left_side, a_shape, b_shape in [
+          (False, (4, 4), (4,)),
           (False, (4, 4), (1, 4,)),
           (False, (3, 3), (4, 3)),
+          (True, (4, 4), (4,)),
           (True, (4, 4), (4, 1)),
           (True, (4, 4), (4, 3)),
           (True, (2, 8, 8), (2, 8, 10)),
@@ -1124,6 +1133,7 @@ class ScipyLinalgTest(jtu.JaxTestCase):
     for dtype in float_types + complex_types
   ))
   def testIssue2131(self, n, dtype):
+    _skip_on_mac_xla_bug()
     args_maker_zeros = lambda: [onp.zeros((n, n), dtype)]
     osp_fun = lambda a: osp.linalg.expm(a)
     jsp_fun = lambda a: jsp.linalg.expm(a)
@@ -1131,6 +1141,35 @@ class ScipyLinalgTest(jtu.JaxTestCase):
                             check_dtypes=True)
     self._CompileAndCheck(jsp_fun, args_maker_zeros, check_dtypes=True)
 
+  @parameterized.named_parameters(jtu.cases_from_list(
+      {"testcase_name": "_lhs={}_rhs={}_lower={}".format(
+          jtu.format_shape_dtype_string(lhs_shape, dtype),
+          jtu.format_shape_dtype_string(rhs_shape, dtype),
+          lower),
+       "lhs_shape": lhs_shape, "rhs_shape": rhs_shape, "dtype": dtype,
+       "rng_factory": rng_factory, "lower": lower}
+      for lhs_shape, rhs_shape in [
+          [(1, 1), (1,)],
+          [(4, 4), (4,)],
+          [(4, 4), (4, 4)],
+      ]
+      for dtype in float_types
+      for lower in [True, False]
+      for rng_factory in [jtu.rand_default]))
+  def testChoSolve(self, lhs_shape, rhs_shape, dtype, lower, rng_factory):
+    rng = rng_factory()
+    _skip_if_unsupported_type(dtype)
+    def args_maker():
+      b = rng(rhs_shape, dtype)
+      if lower:
+        L = onp.tril(rng(lhs_shape, dtype))
+        return [(L, lower), b]
+      else:
+        U = onp.triu(rng(lhs_shape, dtype))
+        return [(U, lower), b]
+    self._CheckAgainstNumpy(osp.linalg.cho_solve, jsp.linalg.cho_solve,
+                            args_maker, check_dtypes=True, tol=1e-3)
+                            
   @parameterized.named_parameters(jtu.cases_from_list(
       {"testcase_name":
        "_lhs={}_rhs={}".format(
